requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.31.0
python-dateutil==2.8.2
twilio==9.3.3
opencv-python==4.10.0.82
mediapipe==0.8.11
scikit-image==0.26.0
numpy==1.26.4
scipy==1.14.1
Pillow==10.4.0
onnxruntime==1.19.2
